# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CMMC Compliance Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence items with filtering
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, needs_classification, pending_review, approved, assessor_ready, rejected, stale, superseded, archived]
        - name: evidenceType
          in: query
          schema:
            type: string
        - name: controlId
          in: query
          schema:
            type: string
        - name: domainId
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: expiringDays
          in: query
          schema:
            type: integer
          description: Filter evidence expiring within N days
      responses:
        "200":
          description: Evidence items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvidenceItem"
...
  /evidence/search:
    get:
      operationId: searchEvidence
      tags: [evidence]
      summary: Full-text search across evidence title, description, tags
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching evidence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvidenceItem"
  /evidence/{id}:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: Get evidence detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evidence item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceDetail"
    patch:
      operationId: updateEvidence
      tags: [evidence]
      summary: Update evidence metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvidenceBody"
      responses:
        "200":
          description: Updated evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
  /evidence/{id}/submit:
    post:
      operationId: submitEvidenceForReview
      tags: [evidence]
      summary: Submit evidence for reviewer approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitEvidenceBody"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
  /evidence/{id}/approve:
    post:
      operationId: approveEvidence
      tags: [evidence]
      summary: Approve evidence (Reviewer/Compliance Manager only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveEvidenceBody"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
  /evidence/{id}/reject:
    post:
      operationId: rejectEvidence
      tags: [evidence]
      summary: Reject evidence with required notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectEvidenceBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
...
  /evidence/{id}/supersede:
    post:
      operationId: supersedeEvidence
      tags: [evidence]
      summary: Supersede old evidence with a new version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupersedeEvidenceBody"
      responses:
        "201":
          description: New version created, old marked superseded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
  /evidence/{id}/link-controls:
    post:
      operationId: linkEvidenceToControls
      tags: [evidence]
      summary: Link evidence to one or more CMMC controls
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkControlsBody"
      responses:
        "200":
          description: Links updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
...
components:
  schemas:
...
    EvidenceItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        evidenceType:
          type: string
          enum: [policy, procedure, screenshot, log, report, ticket, configuration_export, training_record, incident_record, risk_record, approval_record, network_diagram, scan_report, other]
        status:
          type: string
          enum: [draft, needs_classification, pending_review, approved, assessor_ready, rejected, stale, superseded, archived]
        fileName:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        sourceSystem:
          type: string
          nullable: true
        confidentialityLevel:
          type: string
          nullable: true
        version:
          type: string
        tags:
          type: array
          items:
            type: string
        ownerId:
          type: string
        ownerName:
          type: string
        reviewerId:
          type: string
          nullable: true
        reviewerName:
          type: string
          nullable: true
        approverId:
          type: string
          nullable: true
        approverName:
          type: string
          nullable: true
        collectedAt:
          type: string
          format: date-time
          nullable: true
        reviewDueDate:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        linkedControlIds:
          type: array
          items:
            type: string
        linkedControlLabels:
          type: array
          items:
            type: string
        assessorSummary:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - evidenceType
        - status
        - version
        - tags
        - ownerId
        - ownerName
        - linkedControlIds
        - linkedControlLabels
        - createdAt
        - updatedAt
    EvidenceDetail:
      allOf:
        - $ref: "#/components/schemas/EvidenceItem"
        - type: object
          properties:
            internalNotes:
              type: string
              nullable: true
            implementationStatement:
              type: string
              nullable: true
            rejectionNotes:
              type: string
              nullable: true
            previousVersionId:
              type: string
              nullable: true
            auditTrail:
              type: array
              items:
                $ref: "#/components/schemas/AuditLogEntry"
...
    UpdateEvidenceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        evidenceType:
          type: string
        sourceSystem:
          type: string
        confidentialityLevel:
          type: string
        tags:
          type: array
          items:
            type: string
        assessorSummary:
          type: string
        internalNotes:
          type: string
        implementationStatement:
          type: string
        reviewerId:
          type: string
        expiresAt:
          type: string
          format: date-time
        reviewDueDate:
          type: string
          format: date-time
        collectedAt:
          type: string
          format: date-time
    SubmitEvidenceBody:
      type: object
      properties:
        reviewerId:
          type: string
        notes:
          type: string
    ApproveEvidenceBody:
      type: object
      properties:
        assessorSummary:
          type: string
        notes:
          type: string
    RejectEvidenceBody:
      type: object
      properties:
        rejectionNotes:
          type: string
      required:
        - rejectionNotes
    SupersedeEvidenceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        evidenceType:
          type: string
        fileName:
          type: string
        fileData:
          type: string
        notes:
          type: string
      required:
        - title
        - evidenceType
    LinkControlsBody:
      type: object
      properties:
        controlIds:
          type: array
          items:
            type: string
      required:
        - controlIds

    # ─── TASK ─────────────────────────────────────────────────
...
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        entityLabel:
          type: string
          nullable: true
        previousValue:
          type: object
          nullable: true
        newValue:
          type: object
          nullable: true
        ipAddress:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - action
        - entityType
        - entityId
        - timestamp